+++ Database/dbo/Tables/PPMPSupplementaries.sql
@@ -10,15 +10,17 @@
     [Amount]              DECIMAL (19, 5) NOT NULL,
     [Remarks]             NVARCHAR (MAX)  NULL,
     [IsActive]            BIT             NOT NULL,
-    [CreatedByUserId]     BIGINT          NOT NULL,
+    [CreatedByUserId]     INT             NOT NULL,
     [CreatedDate]         DATETIME        NOT NULL,
-    [UpdatedByUserId]     BIGINT          NOT NULL,
+    [UpdatedByUserId]     INT             NOT NULL,
     [UpdatedDate]         DATETIME        NOT NULL,
-    [DeletedByUserId]     BIGINT          NOT NULL,
+    [DeletedByUserId]     INT             NOT NULL,
     [DeletedDate]         DATETIME        NOT NULL,
     [SupplementaryId]     INT             NOT NULL,
     CONSTRAINT [PK__PPMPSupp__6D54EC755C73118F] PRIMARY KEY CLUSTERED ([PPMPSupplementaryId] ASC),
     CONSTRAINT [FKF25AB4FE366CDA2C] FOREIGN KEY ([SupplementaryId]) REFERENCES [dbo].[SS_SupplementaryCatalogue] ([SupplementaryCatalogueId]),
     CONSTRAINT [FKF25AB4FE597D4FBB] FOREIGN KEY ([PPMPId]) REFERENCES [dbo].[PPMPs] ([PPMPId])
 );
+
+
