+++ EF/Models/Ppmpcatalogue.cs
@@ -27,15 +27,15 @@ public partial class Ppmpcatalogue
 
     public bool IsActive { get; set; }
 
-    public long CreatedByUserId { get; set; }
+    public int CreatedByUserId { get; set; }
 
     public DateTime CreatedDate { get; set; }
 
-    public long UpdatedByUserId { get; set; }
+    public int UpdatedByUserId { get; set; }
 
     public DateTime UpdatedDate { get; set; }
 
-    public long DeletedByUserId { get; set; }
+    public int DeletedByUserId { get; set; }
 
     public DateTime DeletedDate { get; set; }
